+++ aftiktuna-rust/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Save file is now removed when reaching the victory/game over frame
 - After victory/game over, the game returns to the main menu
 - "Load Game" button is no longer shown when the save file is absent
+- Made certain action failures more visible, updated some action failure messages
 
 ### Contributors for this release